\name{GeneSetBagging-package}
\alias{GeneSetBagging-package}
\alias{GeneSetBagging}
\docType{package}
\title{
Gene Set Bagging
}
\description{
Performs gene set bagging and calculates the replication probability for every gene set. 
}
\details{
\tabular{ll}{
Package: \tab GeneSetBagging\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-15\cr
License: \tab GNU-2\cr
}

hyperBag() \cr
gseaBag()
}
\author{
Andrew Jaffe \cr
Maintainer: Andrew Jaffe <andrewejaffe@gmail.com>
}
\references{
Jaffe AE, Storey JD, Ji HK, Leek JT. Gene set bagging for estimating the replicability of gene set analyses. Under Review.}
\keyword{}
\seealso{
\code{ limma::geneSetTest, GOstats }
}
\examples{

}
