\name{pExample}
\alias{pExample}
\docType{data}
\title{
Example gene expression dataset 
}

\description{
First 5000 probes with Entrez Gene IDs from  GEO dataset: GSE17913
}
\usage{data(pExample)}
\format{
matrix, genes down rows, samples across columns
}
\details{
}
\source{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE17913}
\references{
}
\examples{
data(pExample)
pExample[1:5,1:5]
}
\keyword{datasets}
